refactor(api): migrate /api/admin/tags
