+++ api/lib/application/tags/index.js
@@ -8,30 +8,6 @@ const register = async function (server) {
   server.route([
     {
       method: 'GET',
-      path: '/api/admin/tags',
-      config: {
-        pre: [
-          {
-            method: (request, h) =>
-              securityPreHandlers.hasAtLeastOneAccessOf([
-                securityPreHandlers.checkAdminMemberHasRoleSuperAdmin,
-                securityPreHandlers.checkAdminMemberHasRoleCertif,
-                securityPreHandlers.checkAdminMemberHasRoleSupport,
-                securityPreHandlers.checkAdminMemberHasRoleMetier,
-              ])(request, h),
-            assign: 'hasAuthorizationToAccessAdminScope',
-          },
-        ],
-        handler: tagController.findAllTags,
-        notes: [
-          "- **Cette route est restreinte aux utilisateurs authentifiés ayant les droits d'accès**\n" +
-            '- Renvoie tous les tags.',
-        ],
-        tags: ['api', 'tags'],
-      },
-    },
-    {
-      method: 'GET',
       path: '/api/admin/tags/{id}/recently-used',
       config: {
         pre: [
