--- api/lib/domain/usecases/find-all-tags.js
@@ -1,5 +1,0 @@
-const findAllTags = function ({ tagRepository }) {
-  return tagRepository.findAll();
-};
-
-export { findAllTags };
