+++ api/src/organizational-entities/application/tag/tag.admin.route.js
@@ -6,6 +6,30 @@ import { tagAdminController } from './tag.admin.controller.js';
 const register = async function (server) {
   server.route([
     {
+      method: 'GET',
+      path: '/api/admin/tags',
+      config: {
+        pre: [
+          {
+            method: (request, h) =>
+              securityPreHandlers.hasAtLeastOneAccessOf([
+                securityPreHandlers.checkAdminMemberHasRoleSuperAdmin,
+                securityPreHandlers.checkAdminMemberHasRoleCertif,
+                securityPreHandlers.checkAdminMemberHasRoleSupport,
+                securityPreHandlers.checkAdminMemberHasRoleMetier,
+              ])(request, h),
+            assign: 'hasAuthorizationToAccessAdminScope',
+          },
+        ],
+        handler: (request, h) => tagAdminController.findAllTags(request, h),
+        notes: [
+          "- **Cette route est restreinte aux utilisateurs authentifiés ayant les droits d'accès**\n" +
+            '- Renvoie tous les tags.',
+        ],
+        tags: ['api', 'tags'],
+      },
+    },
+    {
       method: 'POST',
       path: '/api/admin/tags',
       config: {
