+++ api/tests/organizational-entities/acceptance/application/tag/tag.admin.route.test.js
@@ -6,7 +6,68 @@ import {
   insertUserWithRoleSuperAdmin,
 } from '../../../../test-helper.js';
 
-describe('Acceptance | Organizational Entities | Application | Route | Admin | Organization', function () {
+describe('Acceptance | Organizational Entities | Application | Route | Admin | Tag', function () {
+  describe('GET /api/admin/tags', function () {
+    it('returns a list of tags with 200 HTTP status code', async function () {
+      // given
+      const server = await createServer();
+      const tag1 = databaseBuilder.factory.buildTag({ name: 'TAG1' });
+      const tag2 = databaseBuilder.factory.buildTag({ name: 'TAG2' });
+      await databaseBuilder.commit();
+
+      const userId = (await insertUserWithRoleSuperAdmin()).id;
+
+      const options = {
+        method: 'GET',
+        url: '/api/admin/tags',
+        headers: { authorization: generateValidRequestAuthorizationHeader(userId) },
+      };
+
+      const expectedTags = [
+        {
+          attributes: {
+            name: tag1.name,
+          },
+          id: tag1.id.toString(),
+          type: 'tags',
+        },
+        {
+          attributes: {
+            name: tag2.name,
+          },
+          id: tag2.id.toString(),
+          type: 'tags',
+        },
+      ];
+
+      // when
+      const response = await server.inject(options);
+
+      // then
+      expect(response.statusCode).to.equal(200);
+      expect(response.result.data).to.deep.equal(expectedTags);
+    });
+
+    it('returns 403 HTTP status code when the user authenticated is not SuperAdmin', async function () {
+      // given
+      const server = await createServer();
+      const userId = databaseBuilder.factory.buildUser().id;
+      await databaseBuilder.commit();
+
+      const options = {
+        method: 'GET',
+        url: '/api/admin/tags',
+        headers: { authorization: generateValidRequestAuthorizationHeader(userId) },
+      };
+
+      // when
+      const response = await server.inject(options);
+
+      // then
+      expect(response.statusCode).to.equal(403);
+    });
+  });
+
   describe('POST /api/admin/tags', function () {
     context('when request is valid', function () {
       it('returns the created tag with 201 HTTP status code', async function () {
