+++ api/tests/organizational-entities/integration/application/tag/tag.admin.route.test.js
@@ -1,9 +1,32 @@
 import { tagAdminController } from '../../../../../src/organizational-entities/application/tag/tag.admin.controller.js';
 import * as tagAdminRoutes from '../../../../../src/organizational-entities/application/tag/tag.admin.route.js';
 import { securityPreHandlers } from '../../../../../src/shared/application/security-pre-handlers.js';
-import { expect, HttpTestServer, sinon } from '../../../../test-helper.js';
+import { domainBuilder, expect, HttpTestServer, sinon } from '../../../../test-helper.js';
 
 describe('Integration | Route | Admin | Tag', function () {
+  describe('GET /api/admin/tags', function () {
+    it('returns a response with an http status code OK', async function () {
+      // given
+      const tags = [
+        domainBuilder.buildTag({ name: 'TAG1' }),
+        domainBuilder.buildTag({ name: 'TAG2' }),
+        domainBuilder.buildTag({ name: 'TAG3' }),
+      ];
+
+      sinon.stub(tagAdminController, 'findAllTags').returns(tags);
+      sinon.stub(securityPreHandlers, 'checkAdminMemberHasRoleSuperAdmin').callsFake((request, h) => h.response(true));
+      const httpTestServer = new HttpTestServer();
+      await httpTestServer.register(tagAdminRoutes);
+
+      // when
+      const response = await httpTestServer.request('GET', '/api/admin/tags');
+
+      // then
+      expect(response.statusCode).to.equal(200);
+      sinon.assert.calledOnce(tagAdminController.findAllTags);
+    });
+  });
+
   describe('POST /api/admin/tags', function () {
     it('returns forbidden access if admin member has CERTIF or SUPPORT or METIER role', async function () {
       // given
