+++ api/tests/organizational-entities/unit/domain/usecases/find-all-tags.usecase.test.js
@@ -1,7 +1,7 @@
-import { findAllTags } from '../../../../lib/domain/usecases/find-all-tags.js';
-import { domainBuilder, expect, sinon } from '../../../test-helper.js';
+import { findAllTags } from '../../../../../src/organizational-entities/domain/usecases/find-all-tags.usecase.js';
+import { domainBuilder, expect, sinon } from '../../../../test-helper.js';
 
-describe('Unit | UseCase | find-all-tags', function () {
+describe('Unit | Organizational Entities | Domain | UseCase | find-all-tags', function () {
   let tagRepository;
 
   beforeEach(function () {
@@ -10,7 +10,7 @@ describe('Unit | UseCase | find-all-tags', function () {
     };
   });
 
-  it('should return tags', async function () {
+  it('returns tags', async function () {
     // given
     const tags = [
       domainBuilder.buildTag({ name: 'TAG1' }),
@@ -27,7 +27,7 @@ describe('Unit | UseCase | find-all-tags', function () {
     expect(tagRepository.findAll).to.have.been.calledOnce;
   });
 
-  it('should return an empty array when no tags found', async function () {
+  it('returns an empty array when no tags found', async function () {
     // given
     tagRepository.findAll.returns([]);
