--- api/tests/unit/application/tags/index_test.js
@@ -1,29 +1,0 @@
-import * as moduleUnderTest from '../../../../lib/application/tags/index.js';
-import { tagController } from '../../../../lib/application/tags/tag-controller.js';
-import { securityPreHandlers } from '../../../../src/shared/application/security-pre-handlers.js';
-import { domainBuilder, expect, HttpTestServer, sinon } from '../../../test-helper.js';
-
-describe('Unit | Application | Router | tag-router', function () {
-  describe('GET /api/admin/tags', function () {
-    it('return a response with an http status code OK', async function () {
-      // given
-      const tags = [
-        domainBuilder.buildTag({ name: 'TAG1' }),
-        domainBuilder.buildTag({ name: 'TAG2' }),
-        domainBuilder.buildTag({ name: 'TAG3' }),
-      ];
-
-      sinon.stub(tagController, 'findAllTags').returns(tags);
-      sinon.stub(securityPreHandlers, 'checkAdminMemberHasRoleSuperAdmin').callsFake((request, h) => h.response(true));
-      const httpTestServer = new HttpTestServer();
-      await httpTestServer.register(moduleUnderTest);
-
-      // when
-      const response = await httpTestServer.request('GET', '/api/admin/tags');
-
-      // then
-      expect(response.statusCode).to.equal(200);
-      sinon.assert.calledOnce(tagController.findAllTags);
-    });
-  });
-});
